A PagingSource custom implementation for Room

+++ data/src/test/java/ua/in/khol/oleh/githobbit/data/network/github/RepoFactory.kt
@@ -1,10 +1,10 @@
 package ua.`in`.khol.oleh.githobbit.data.network.github
 
-import ua.`in`.khol.oleh.githobbit.domain.entity.Repo
+import ua.`in`.khol.oleh.githobbit.domain.model.Repo
 
 class RepoFactory {
     fun create(id: Int): Repo = Repo(
-        id = id,
+        id = id.toLong(),
         ownerName = "Owner $id",
         ownerImage = "",
         repoName = "Name $id",
